requirements (1).txt
fastapi==0.103.1
uvicorn==0.23.2
pinecone-client==2.2.2
openai==1.2.0
langchain==0.0.300
PyPDF2==3.0.1
scikit-learn==1.3.1
nltk==3.8.1
python-dotenv==1.0.0